+++ packages/_example/src/forest/customizations/account.ts
@@ -2,5 +2,9 @@ import { CollectionCustomizer } from '@forestadmin/agent';
 
 import { Schema } from '../typings';
 
-export default (collection: CollectionCustomizer<Schema, 'account'>) =>
+export default (collection: CollectionCustomizer<Schema, 'account'>) => {
+  collection.importField('address_city', { path: 'address:city' });
+  collection.importField('address_country', { path: 'address:country' });
+
   collection.addManyToOneRelation('store', 'store', { foreignKey: 'storeId' });
+};
