+++ packages/datasource-customizer/src/decorators/write/datasource.ts
@@ -1,34 +1,15 @@
-import { Caller, Collection, DataSource, DataSourceSchema } from '@forestadmin/datasource-toolkit';
+import { DataSource } from '@forestadmin/datasource-toolkit';
 
 import DataSourceDecorator from '../datasource-decorator';
-import RelationWriterCollectionDecorator from './collection-relation-handler';
-import WriteReplacerCollectionDecorator from './collection-write-replace';
+import RelationWriterCollectionDecorator from './relation/collection';
+import WriteReplacerCollectionDecorator from './write-replace/collection';
 
-export default class WriteDecorator implements DataSource {
-  private readonly dataSource: DataSource;
-
+/**
+ * Decorator which allows to change the behavior of the write actions for given fields.
+ */
+export default class WriteDataSourceDecorator extends DataSourceDecorator<WriteReplacerCollectionDecorator> {
   constructor(childDataSource: DataSource) {
-    const decorator = new DataSourceDecorator(childDataSource, WriteReplacerCollectionDecorator);
-    this.dataSource = new DataSourceDecorator(decorator, RelationWriterCollectionDecorator);
-  }
-
-  get collections(): Collection[] {
-    return this.dataSource.collections;
-  }
-
-  get schema(): DataSourceSchema {
-    return this.dataSource.schema;
-  }
-
-  getCollection(name: string): Collection {
-    return this.dataSource.getCollection(name);
-  }
-
-  addCollection(collection: Collection): void {
-    return this.dataSource.addCollection(collection);
-  }
-
-  renderChart(caller: Caller, name: string): Promise<unknown> {
-    return this.dataSource.renderChart(caller, name);
+    const decorator = new DataSourceDecorator(childDataSource, RelationWriterCollectionDecorator);
+    super(decorator, WriteReplacerCollectionDecorator);
   }
 }
