+++ packages/datasource-customizer/src/decorators/write/write-replace/context.ts
@@ -1,7 +1,7 @@
 import { Caller, Collection } from '@forestadmin/datasource-toolkit';
 
-import CollectionCustomizationContext from '../../context/collection-context';
-import { TCollectionName, TPartialSimpleRow, TSchema } from '../../templates';
+import CollectionCustomizationContext from '../../../context/collection-context';
+import { TCollectionName, TPartialSimpleRow, TSchema } from '../../../templates';
 
 export default class WriteCustomizationContext<
   S extends TSchema = TSchema,
