+++ packages/datasource-customizer/src/index.ts
@@ -12,7 +12,7 @@ export { OperatorDefinition } from './decorators/operators-emulate/types';
 export { RelationDefinition } from './decorators/relation/types';
 export { SearchDefinition } from './decorators/search/types';
 export { SegmentDefinition } from './decorators/segment/types';
-export { WriteDefinition } from './decorators/write/types';
+export { WriteDefinition } from './decorators/write/write-replace/types';
 export * from './decorators/hook/types';
 
 // Context
